First proper working sqlx query in Rust
1. modified DATABASE_URL in .env file to include password to the url string
2. mysql.session.sql added for who knows what reason
3. pyfi_db.sql modified to change image from NOT NULL to a nullable type
4. changes to the sqlx main file:
a. changed the name of password field of UserAccount struct to user_password
b. changed the date to include double quotes
c. changed comments to update the mysql database url and added a "unique" comment annotation to username field of UserAccount struct
d. used dotenvy::dotenv() to make connecting to a database easier