Ladies and gentlemen, butterfingersd-verify is now 100% operational!
1. overhauled butterfingersd-verify so that:
a. the initializing phase of the program is put *outside* of the loop
b. print parameter in match_cb callback is used to assign the matched fingerprint's username
c. attendance is only recorded when the fingerprint has a corresponding uuid (which only happens when it gets a match)
d. temporarily removed the close_sync function at the bottom of main.rs, not entirely sure if it is good to do so but it ensures that the looping works as intended
e. a lot of debug and print statements: used to show that the program runs as expected

2. updated the crates via cargo update as per usual