wide reaching pool update
1. MySqlPool is now created at the start, in a state variable of type ManagedMySqlPool
This has three implications:
a. you don't need to call query_db anymore and now it only needs to be called once.
b. creating the pool is no longer a matter of concern when it comes to overhead as creating the pool is also only called once.
c. any MySql query now only needs to access the state variable to be able to access the pool and to execute it.
2. managed parameter for verify_fingerprint temporarily set to _managed (as it is currently unused)
3. cargo fmt
4. temporarily commented out *every* method call of pool.close().await
(As apparently you don't need to do that when pool sharing)